Fix shaded resize base, cycle menu stacking
Removed the per-render `STACK_MODE_ABOVE` raise from
`desktop_render_clients`, which raised all managed windows on every
repaint pass and buried the cycle menu.  Added `is_outdated=true`
after every titlebar button click (was already done for scroll-wheel
and double-click but not for the button-click path).

+++ include/client.h
@@ -709,7 +709,7 @@ int client_send_event_set_icon(client_td *client, const char *icon_name);
  * @note Complexity: @e O(1)
  */
 #define client_is_iconified(w) \
-    ((w)->properties.state & CLIENT_STATE_ICONIFIED)
+    ((w)->properties.state & (uint16_t) CLIENT_STATE_ICONIFIED)
 
 /**
  * @brief Macro that evaluates to the client maximization state
